requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
anthropic>=0.3.0
websockets>=10.0
redis>=4.0.0
sqlalchemy>=1.4.0
pandas>=1.3.0
numpy>=1.21.0
plotly>=5.3.0
python-dotenv>=0.19.0
pyyaml>=5.4.0
pytest>=6.2.0
httpx>=0.23.0
mcp-python>=0.1.0  # Add MCP Python SDK
asyncio>=3.4.3
aiohttp>=3.8.0
